improving the emitted code for jump instructions

+++ objeck-lang/src/compiler/tree.h
@@ -216,7 +216,7 @@ namespace frontend {
     }
 
     ScopeTable* GetNextChild() {
-      if(child_pos < children.size()) {
+      if(child_pos < (int)children.size()) {
 	return children[child_pos++];
       }
